+++ Frameworks/MABookmarksBar/MABookmarksBar/MABookmarksBarItem.swift
@@ -46,4 +46,11 @@ open class MABookmarksBarItem: NSButton {
     public required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+
+    open func updateColors(background: NSColor, titleColor: NSColor) {
+        wantsLayer = true
+        tabTitle.textColor = titleColor
+        layer?.backgroundColor = background.cgColor
+        (cell as? NSButtonCell)?.backgroundColor = background
+    }
 }