+++ Changes.md
@@ -1,6 +1,8 @@
 #  Changes
 
 ## v1.2
+- Few improvements with a few stuff
+- Added Support for colors in the bookmarks bar
 - Created a basic bookmarks bar
 - Fixed problem where removing the webView's observer would cause an error
 - Fixed error where being on a tab and trying to close another tab, will not close the webview