+++ Frameworks/MABookmarksBar/Changes.md
@@ -1,4 +1,5 @@
 # Changes
 
 ## v1.0
+- Added Support for colors in the bookmarks bar
 - Created a basic bookmarks bar